INGC_CRM/requirements.txt
Django==4.1.2
djangorestframework==3.14.0
django-cors-headers==3.13.0
qrcode==7.3.1
Pillow==9.2.0
PyJWT==2.5.0
colorama==0.4.6
flask==2.2.2
